lab5/requirements.txt
PyYAML==6.0.3
xmltodict==1.0.2